requirements.txt
apache-airflow==2.7.0
cassandra_driver==3.28.0
kafka_python==2.0.2
pyspark==3.5.0
pytz==2024.1
Requests==2.31.0
